+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # CHANGELOG workflo/macros
 
+## Unreleased
+
+### Changed
+
+* `workflo.macros.command/resolve-command` now returns the
+  command definition map instead of its qualified symbol.
+
+### Added
+
+* `workflo.macros.command/resolve-command-sym` has been added
+  to provide the old functionality of `resolve-command`.
+
 ## 0.2.10
 
-## Changed
+### Changed
 
 * Queries now support joins in combination with links, that is,
   links that are joined with models, properties or recursion, e.g.:
@@ -15,14 +27,14 @@
     :join-target [{:name db/id :type :property}
                   {:name user/name :type :property}]}`.
 
-## Fixed
+### Fixed
 
 * Generate '_ for global links in Om Next queries, not ''_.
 * The CLJS generator for the `::map-destructuring-keys` spec now works.
 
 ## 0.2.9
 
-## Added
+### Added
 
 * New functions to resolve variables and variable paths using
   potentially deeply nested parameter/value maps:
@@ -34,7 +46,7 @@
     - `workflo.macros.query.bind/resolve-path`
     - `workflo.macros.query.bind/resolve`
 
-## Fixed
+### Fixed
 
 * When binding query parameters, only recurse into join targets
   if they are real subqueries.